Move treebank_to_wordnet function to utils
